Implement notion of shader binding table entries
An "entry" in the shader binding table may consist of one or more shader
groups. A typical use case for this is to create an entry with multiple
miss shaders. The ray generation shader chooses dynamically from one of
the shaders in the entry bound with the call to vkCmdTraceRaysKHR.