Make elements able to receive events
Add a generic EventListenerRegistryBase abstraction. The initial idea
was to extract this basic abstraction from vkb::EventHandler and add it
to vk_base. This turned out to be unfeasible however as
vkb::EventHandler requires its properties and methods to be static,
which should not be the case for ui::Element.