Store matrix ID in DrawableData instead of Node*
To accomplish this I needed to change the way a Node's global matrix is
calculated. I introduced a virtual call which is called every time the
local transformation changes because that's just very convenient. Thus,
I don't need logic in Node::getGlobalTransform anymore and can fully
rely on the global matrix's ID.