Implement support for multiple viewports
Add a RenderTarget class. Make RenderConfig depend on RenderTargets
instead of directly on the swapchain.

We can create multiple render configs. Each render config has a render
target to which it draws (multiple configs can draw to the same target)
and a viewport, which determines to which area on the render target the
image will be drawn.

Viewports are not real objects, but rather a setting on RenderConfig.
Because every render config has a single viewport (and a viewport has
always a render config associated with it), the function is simply
`RenderConfig::setViewport(offset, size)`.

On swapchain recreate, the render target object has to be recreated
because the contained images and image views no longer exist, and
RenderConfig::setRenderTarget must be called again.

Renderer now accepts multiple DrawConfig structs, where each struct
defines a viewport.

Because some descriptors will be specific to a single viewport/render
config, a few changes had to be made to the DescriptorRegistry and to
the way pipeline layouts store descriptors. A descriptor registry
template instantiation for a specific render config type will now assign
unique IDs to descriptor names and store them statically. An instance of
the descriptor registry object will be stored in each render config
object and serve as a storage of descriptor providers; analogously to
the PipelineStorage. Descriptors are now queried from the render config
during command recording, much like pipelines. That's why we have IDs
for fast runtime lookup and names for user-friendly lookup at layout
creation time. Therefore, pipeline layouts can now store descriptor
providers as before (one can still specify specialized descriptors for
locally used pipeline layouts; an example is the final lighting pass's
pipeline), but also IDs, which are more like descriptor layouts that
have different instantiations for their sets for each render config.

This descriptor mechanic looks complex, and sadly, it is, but the entire
name-to-ID-mapping magic happens entirely behind the scenes.

Various small improvements to the deferred shading pipeline:

 - Improve barriers in the g-buffer pass

 - Add a getMouseWorldPos function to DeferredRenderConfig so that the
   deferred pass does not have to be accessed directly.

 - Make final lighting output dimensions variable

 - Move render graph argument from DeferredRenderConfig's constructor to
   the create info

 - Fix the maximum pixel coordinates in final_lighting.comp

Other minor fixes:

 - Check for Instance::hasRayTracing before setting the deviceAddress
   flags on buffers in AssetRegistry

 - Fix segfault in test/font.cpp