Implement better polling mechanism for EventThread
Use a condition variable to put the thread to sleep until an event is
ready for processing, instead of the brute-force `while(!ready);` loop
that was used before.

Also add a note about possible deadlocks to ThreadPool::async and unlock
a mutex before notifying the condition variable.

Fix waiting for workers to become idle in ThreadPool.