Add my Table stuff as a component library
Implement various small improvements for the existing component
infrastructure.

Also add a default implementation for a typesafe component ID that can
be used in ComponentStorage. It can be instantiated with type tags in
using-declarations to define different types of IDs.

Add torch_util dependency to componentlib (for Maybe<>).

Move unique storage declaration to TableTraits<>. The TableTraits
template can be specialized with a type definition `UniqueStorage` to
declare that a type shall be stored as unique_ptrs in the table.