Add unified TextureData structure
TextureData contains raw pixel data. Textures will be imported to this
format.

Add a loadTexture function (equivalent to loadGeometry) that loads image
data into a TextureData structure.

Use TextureData to load textures into the asset registry. vkb::Image was
used previously, but the data-based approach is consistent with the
approach to geometries. Also, it is much easier to create device
resources *inside* the asset registry.