Make AssetReference constructible from AssetSource
The makes for way more convenient use in non-asset-path-based asset
creation, for example during import of external formats.

Add AssetManager::create function that creates an asset from an
AssetSource.

Relatedly, implement a mechanism that resolves references of imported
assets: The AssetManager creates a wrapper around AssetSource (which is
the most basic form in which assets are created) that resolves
references of assets at the time they are loaded in their corresponding
asset registry module.