Add TargetType property to shaders
Defines the output format, i.e. GLSL or SPIRV. This is an optional
property, the default can be set via the --spv command-line option. The
property, if set, will override the default value.

This places the 'include-only shader' hack on a more solid ground,
though the *default* value for .glsl shaders is still being handled
differently.