Implement an almost-working windows build
The build still has two linker errors when linking the torch library. It
seems to have something to do with protobuf, though I can't reproduce it
in a minimal example and have generally do idea what to do here.

I still commit this almost-complete thing because it improved some
CMake- and well as c++-code, and I don't want to rebase this again after
I've done some work on the actual project.