Extend material-fragment shader to generic shaders
Create a proper mechanism to specify fragment shader inputs (from the
vertex shader) by specifying single locations as resources, without any
custom code.

Use this generic mechanism to move the syntax to a more generic
description:

 - Rename VertexOutput to generic ShaderInput

 - Rename fragmentOutput* to shaderOutput*

because it is now possible to use this mechanism for all kinds of
shaders.