Add a code-include mechanism to module builder
Add file includes to ShaderModuleBuilder that replace the post-resource
include hack in ShaderCapabilityConfig. These new includes also allow
setting replacement variables to capability accessors.

Add an `inline` flag to ShaderCodeCompiler that controls whether all
expressions are inlined or written to intermediate variables. This is
required for the aforementioned variable replacements.