Use new system as MaterialData
Improve (maybe, not so sure yet) the material specialization mechanism.
Try to streamline the material creation process and data types it uses.

Add SimpleMaterialData to replace the old MaterialData struct. Use this
in imprt/export functions for now, just to get it to compile. I might
even implement this simpler mechanism.

Create the required ShaderDescriptorConfig in AssetRegistry's
constructor.

Also modify Drawable to use the new MaterialRuntime.