Various fixes for material and shader system
Add an optional type annotation to code::ValueT. Used to give a type to
expressions for which a type cannot be determined automatically and make
them assignable to a temporary value.

Fix an assert in MaterialRuntime::pushConstant.

Require GL_EXT_nonuniform_qualifier for shadow map descriptor resource.

Implement a provider for the vertexNormal fragment capability in the
vertex shader.

Add an additional module settings structure to ShaderModuleBuilder that
holds a set of global settings for a shader module. Compile it to code
in the ShaderModuleCompiler. The first and only available setting is
early fragment tests.