Add AssetStorage to manage asset data
Use AssetStorage, which provides a high-level interface to store and
retrieve asset data and the corresponding metadata, in the AssetManager
to load assets. Remove AssetPathSource. AssetStorage allows much better
isolation of concerns in the data loading process.

AssetStorage uses the new DataStorage interface. DataStorage acts as a
very simple key-value store. It is used to abstract access to asset data
from the physical storage technology. Also FilesystemDataStorage as the
first implementation of the interface.

Make the structure AssetMetadata more important. Add a virtual method
AssetSource<>::getMetadata to replace getAssetName.

Implement a dynamic representation of asset types: AssetType. Add an
AssetType field to AssetMetadata.