Move asset device stuff to AssetDescriptor
Simplify AssetRegistry. It is now only an interface to the modules.

Implement all of Torch's asset descriptors in the AssetDescriptor class.
This is more consistent with the general descriptor convention,
encapsulates the messy device stuff well, and lets me separate device
resources better from the general asset management (AssetManager and
related things) architecture.