Add unit tests for AssetManager
Add the define_asset_type.h header which defines a utility macro for
easy asset type definition in tests.

Fix bugs discovered in the tests and modify the new AssetManager
interface slightly to be more consistent (well, hopefully).

Because the AssetReference resolution was moved to AssetManager, we can
now use that in ManagerTraits and take an AssetManager object with all
AssetPath functionality. As it should be.