Refactor VulkanDebug and VulkanInstance
Remove the statically initialized members from VulkanDebug and create
them when the VulkanDebug object is created. Use the new Logger
interface to log Vulkan debug messages.

Add more configuration parameters to VulkanInstance creation and store
them all in a VulkanInstanceCreateInfo struct. Fix the
getRequiredInstanceExtensions function in the case that surface creation
is not supported.