Simplify ShaderOutputNode to ShaderOutputInterface
ShaderOutputNode was way too complicated. The new ShaderOutputInterface
can declare assignments or function calls of arbitrary code::Values
(both of which are statements) that are then appended to the shader's
main function in the ShaderModuleCompiler.

This is now the preferred way to create statements. One should not
modify a shader's main function manually.