Implement rchit shader generation from code graph
To achieve this, I had to decouple the capability config from the module
builder. One now specifies a capability configuration for individual
compilations of shader code graphs, not for entire module builders. This
allows me to compile the same node graph with different capability
configurations (e.g. one for fragment shaders in deferred shading and
another for closest hit shaders in ray tracing).

This is used in the FragmentModule class, which can now generate a
fragment shader and/or a closest-hit shader from the same material
definition.

Sadly, I also had to implement capability accesses and runtime constant
(read: specialization constant) definitions as shader code values.
Originally, I wanted to have resources decoupled from 'pure' shader
code, but this new solution is the best one I could find to decouple
creation of capability accesses (i.e. the module builder) from the
definition of the shader code which implements the capabilities. An
upside to this is that it reduces persistent state in the code/module
builder and moves it to a short-lived computation in the
ShaderModuleCompiler.

Adapt the existing (hand-written) ray shaders to the future format of
the auto-generated shaders.

Apply some additional refactorings:

 - Always add type annotations to cast expressions. Constructors also
   count as cast expressions.

 - De-duplicate runtime constant requests.

 - Make code primitive references (`code::Value`) shared pointers. This
   makes the entire thing a lot safer and removes some additional
bookkeeping.