Implement StableTable backend
Add `StableTableImpl` with the corresponding iterators as a second backend
for `Table`. It is a pool-like implementation which keeps memory
locations stable by managing them in chunks.

This replaces the `TableTraits<>::UniqueStorage` hack. Remove the entire
`TableTraits` template.

Move the unit tests to a common header with plug-and-play functionality
for the different table backends; the same tests are executed for
IndirectTable and StableTable.

Also fix some bugs in the join iterator during the process.