Fix pipeline compiler skipping some shader sources
The pipeline compiler was not able to include shader sources defined as
inline object references in the generated shader database and therefore
did not compile those sources to SPIRV.

A new global list of all shader modules in the pipeline result fixes
this. Also it makes shader database generation cleaner.