[medit] Move controls to state-machine architecture
We now have a stack state machine, but we still try to keep the states
themselves as free of internal state as possible. We try to achieve this
by passing all necessary input and output data to the state's `update`
method. This also allows us to provide controlled interfaces to the
states through which they can create output.