Make input processing in Swapchain configurable
Add an `InputProcessor` interface. `Swapchain` calls its functions when
a user input event happens at the underlying window.

The input processor implementation `InputEventSpawner` spawns events at
the `EventHandler`, which is what the Swapchain's job was until now.
`Window` sets this input processor as a default option if no custom
implementation is configured.

As we can no longer rely on the previously-default event system to be
enabled at all, we remove the swapchain recreation-related events
entirely and add the possibility to register callbacks directly at the
swapchain instead.