Implement complex stage ordering in RenderGraph
Rewrite RenderGraph completely such that it now allows us to specify
complex ordering among stages and iterate over them in that order. This
will be used to implement the new barrier system.

Also remove the render stage member from TaskEnvironment and pass the
stage to the raster render pass on creation.