Make shadows per-scene, not per-viewports resources
Remove the ShadowPool from the public interface, use it privately in the
RasterPlugin on a per-scene basis. Hook into the LightSceneModule to
create device resources for shadows when needed. The LightSceneModule
only contains *descriptions* of shadow maps and the RasterPlugin is
responsible for creating actual images and render passes for them.

Rework the interface to lights a bit in the process.