Don't compile SPIR-V code in `linkShaderProgram`
Move SPIR-V compilation of generated GLSL code from `linkShaderProgram`
to `MaterialProgram::MaterialProgram`.

We keep the entire shader builder library free from SPIR-V, which should
be an implementation detail. We move the heavy load to the later stage
(creating the material program), where it can be better anticipated and
handled. Additionally, this paves the way to implement a direct SPIR-V
generator. Some day.