Pass AssetManager to asset modules
Move AssetRegistry (the device data registry) more into the background,
and instead pass the asset manager around. There is no reason that asset
registries shouldn't be able to use the full power of the asset manager
to use other asset types properly.

Also improve the ordering of the default asset modules during
registration and destruction.

Doing this, also remove AssetManagerBase from the public interface.
Instead of inheriting from it, make it the 'implementation' component of
AssetManager. ... it's just always better this way.