Always specify a file name to glslc
Passing a null filename to glslc (as in the case of generated shaders)
results in a failed assertion in the glslc code. This is a bug in my
opinion, but we still have to work with it. Use an artifical file name
that also signifies that the shader is an auto-generated material
shader.